\name{lmenssp-package}
\alias{lmenssp-package}
\alias{lmenssp-package}
\docType{package}
\title{
Linear Mixed Effects Models with Non-stationary Stochastic Processes
}
\description{
Obtains maximum likelihood estimates of the model parameters, filters, smooths 
and forecasts random components of the model for the following 
processes: 1) Brownian motion, 2) integrated Brownian motion, 3) integrated 
Ornstein-Uhlenbeck process, 4) stationary process with powered correlation 
function, 5) stationary process with Matern correlation function, 
under multivariate normal and t response distributions. It also contains 
miscellaneous functions for diagnostic checks, boostrap standard error calculation, etc. 
}
\details{
\tabular{ll}{
Package: \tab lmenssp\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2016-07-23\cr
License: \tab GPL (>=2) \cr
}
}
\references{

Asar O, Ritchie J, Kalra P, Diggle PJ (2015) Acute kidney injury amongst chronic kidney disease 
patients: a case-study in statistical modelling. To be submitted.

Diggle PJ (1988) An approach to the analysis of repeated measurements. \emph{Biometrics}, \bold{44}, 959-971.

Diggle PJ, Heagerty PJ, Liang K-Y, Zeger SL. (2002) \emph{Analysis of Longitudinal Data, 2nd edition}. 
Oxford University Press: Oxford.

Diggle PJ, Ribeiro PJ Jr. (2007) \emph{Model-based Geostatistics}. Springer-Verlag: New York.

Diggle PJ, Sousa I, Asar O (2015) Real time monitoring of progression towards renal failure in primary care patients.
\emph{Biostatistics}, 16(3), 522-536. 

Laird NM, Ware JH (1982) Random-effects models for longitudinal data. \emph{Biometrics}, \bold{38}, 134-147.

Matern B. (1960) \emph{Spatial Variation}. Statens Skogsforsningsinstitut, Stockholm.

Pinheiro JC, Liu C, Wu YN. (2001) Efficient algorithms for robust estimation in linear mixed-effects 
models using the multivariate t distribution. \emph{Journal of Computational and Graphical Statistics} \bold{10}, 249-276.

Ross SM (1996) \emph{Stochastic processes}. John Wiley & Sons, New Jersey.

Taylor JMG, Cumberland WG, Sy JP (1994) A Stochastic Model for Analysis of Longitudinal AIDS Data. 
\emph{Journal of the American Statistical Association}, \bold{89}, 727-736.
}

